% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modana-package.R
\docType{package}
\name{Modana-package}
\alias{Modana}
\alias{Modana-package}
\title{Modana: Refined Moderation Analysis with binary outcomes}
\description{
This package employs the symmetry property concerning the ratio of odds ratios, which suggests that both the main treatment and the heterogeneous treatment effects could be equivalently estimated via a role exchange between the outcome and treatment variable in logistic regression models. Refined inference on moderating effects is obtained by rearranging data and combing two models into one via a generalized estimating equation (GEE) approach.
}
\keyword{internal}
