% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderation.R
\name{refinedmod}
\alias{refinedmod}
\alias{swaparr}
\alias{sim_data}
\alias{summary.refinedmod}
\alias{print.refinedmod}
\alias{plot.refined}
\title{Refined moderation Analysis with binary outcomes}
\usage{
refinedmod(
  formula,
  y = "y",
  trt = "trt",
  effmod = NULL,
  data,
  detail = FALSE,
  ...
)

swaparr(data, y = "y", trt = "trt")

sim_data(
  n = 100,
  b0,
  a0 = NULL,
  link.function = "logistic",
  sigma = 1,
  uniform = TRUE,
  c0 = 1,
  binary.Xs = FALSE,
  rho = 0.2,
  observational = FALSE,
  trt.p = 0.5,
  interaction = NULL,
  details = F
)

\method{summary}{refinedmod}(object, ...)

\method{print}{refinedmod}(x, ...)

\method{plot}{refined}(x, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{y}{name of a binary response column in data.}

\item{trt}{name of a binary treatment column in data.}

\item{effmod}{a vector of potential effect modifiers used in the data.}

\item{data}{original data frame containing the variables in the model.}

\item{detail}{a logical argument specified to print summary results of the direct and inverse models}

\item{...}{other argument not in use at the moment}

\item{n}{the number of observations to generate.}

\item{b0}{a vector of coefficients for the linear predictor in the outcome model. The length of \code{b0} should be equal to the number of covariates plus one (for the intercept term). The first element of \code{b0} corresponds to the intercept, and the remaining elements correspond to the coefficients for the covariates which may include include interaction coefficients.}

\item{a0}{a vector of coefficients for the linear predictor in the treatment assignment model. The length of \code{a0} should be equal to the number of covariates plus one (for the intercept term). The first element of a0 corresponds to the intercept, and the remaining elements correspond to the coefficients for the covariates. This argument is only used if `observational = TRUE`.}

\item{link.function}{a character string specifying the link function to use in the outcome model. The default is `logistic`, which corresponds to the logistic link function. The other option is `exp`, which corresponds to the log-linear link function.}

\item{sigma}{the standard deviation of the covariate distribution. The default value is `1`.}

\item{uniform}{a logical value indicating whether to transform the covariates to a uniform distribution on the interval `[0, c0]`. The default value is `FALSE`.}

\item{c0}{the upper limit of the uniform distribution used to transform the covariates. The default value is 1.}

\item{binary.Xs}{a logical value indicating whether to change some columns of continuous covariates into binary variables. The default value is `FALSE`.}

\item{rho}{the correlation parameter used to generate the covariance matrix. The default value is `0.2`.}

\item{observational}{a logical value indicating whether to generate data for an observational study. The default value is FALSE.}

\item{trt.p}{the probability of receiving the treatment. This argument is only used if `observational = FALSE`. The default value is 0.5.}

\item{interaction}{a vector of indices indicating which covariates should be used to create interaction terms with the treatment variable. The length of interaction should be less than or equal to the number of covariates. The default value is `NULL`, which indicates no interaction terms should be used.}

\item{details}{a logical value indicating whether to print details about the generated data. The default value is FALSE.}

\item{object}{an object class of \code{refinedmod}}

\item{x}{an object class of \code{refinedmod}}
}
\value{
An object of type \code{glm} and \code{geeglm}.

New data frame created by swapping selected columns \code{y} and \code{trt} from \code{data} using the sequence of indices constructed mimicking a longitudinal with 2 clusters.
}
\description{
This package employs the symmetry property concerning the ratio of odds ratios, which suggests that heterogeneous treatment effects could
be equivalently estimated via a role exchange between the outcome and treatment variable in logistic regression
models. Refined inference on moderating effects is obtained by rearranging data and combing two models into
one via a generalized estimating equation (GEE) approach.
}
\details{
Three models of interest are fitted. The generalized linear models \code{\link[stats:glm]{glm}} (for the direct and inverse models) are fitted to verify the symmetry property of OR/RR in moderation analysis for the main treatment effect as well as the moderating effects.
The main idea of this computing trick is to rearrange data into a format that micmicks a clustered or longitudinal data. A clustered data of size `2` is obtained simply by swapping the roles of the response variable \code{y} and treatment variable \code{trt}
columns in the original data \code{data} and then stacking the resultant data on top of the original \code{data}. The dummy variable \code{z}  is introduced to distinguish the two glm models.

swaps the columns specified by \code{y} and \code{trt}, appends the swapped data frame to the original data frame \code{data}, and adds two new columns \code{(group and ID)} to the merged data frame

The function \code{sim_data()} generates simulated data for a randomized controlled trial or an observational study
}
\examples{
set.seed(1001)
b0 <- c(-1.5, 1.5, 1.5, -2, 2, -.5, 2, 1.5)
a0 <- c(-1, 1.5, -1, 1.5)
datt <- sim_data(n = 100, b0, a0 = NULL, binary.Xs = FALSE,
                  sigma = 1, uniform = FALSE, c0 = 1,
                   link.function = "logistic", rho = 0.2,
                   observational = FALSE, trt.p = 0.5,
                   interaction = 1:2, details = FALSE)

getres <- refinedmod(formula = y ~ trt + x1 + x2 + x3 + x4,
                     detail = FALSE, y = "y",
                    trt = "trt", data = datt,
                     effmod = c("x1", "x2"),
                      corstr = "independence")
names(getres)
print(getres$Refinedmodel)
summary(getres$Refinedmodel)
summary(getres$Directmodel)
#plot(getres$Directmodel)
}
\references{
\href{Anto E, Su X. Refined moderation analysis with binary outcomes in precision medicine research. Statistical Methods in Medical Research. 2023;0(0). doi:10.1177/09622802231151206}{Anto E, Su X. Refined moderation analysis with binary outcomes in precision medicine research. Statistical Methods in Medical Research. 2023;0(0). doi:10.1177/09622802231151206}
}
