% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimData}
\alias{SimData}
\title{SimData data}
\format{
An object of class \code{refinedmod}
}
\usage{
data(SimData)
}
\arguments{
\item{SimData}{The data for moderation analysis}
}
\description{
Dataset was simulated from 100 iid subjects with 4 covariates that were randomly assigned to a binary treatment. 
A binary outcome response variable was generated involving the treatment by covariate interactions of (x1, and x2)
}
\examples{
head(data(SimData))
}
\keyword{datasets}
