% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Warts}
\alias{Warts}
\title{A dataset comparing immunotherapy to cryotherapy treatments and their effeteness of removing warts}
\format{
An object of class \code{refinedmod}
}
\usage{
data(Warts)
}
\arguments{
\item{Warts}{The data for moderation analysis}
}
\description{
A dataset comparing immunotherapy to cryotherapy treatments and their effeteness of removing warts
}
\examples{
head(data(Warts))
}
\keyword{datasets}
